+++ src/weather/models/MapInfoModel.cpp
@@ -1,6 +1,6 @@
 /*
 * Vremenar
-* Copyright (C) 2021 Tadej Novak <tadej@tano.si>
+* Copyright (C) 2022 Tadej Novak <tadej@tano.si>
 *
 * This application is bi-licensed under the GNU General Public License
 * Version 3 or later as well as Mozilla Public License Version 2.
@@ -49,7 +49,7 @@ QStringList MapInfoModel::list() const
 std::vector<Weather::MapType> MapInfoModel::types() const
 {
     std::vector<Weather::MapType> l;
-    l.reserve(rowCount());
+    l.reserve(static_cast<size_t>(rowCount()));
     for (int i = 0; i < rowCount(); i++) {
         l.push_back(row<MapInfo>(i)->type());
     }
@@ -57,3 +57,5 @@ std::vector<Weather::MapType> MapInfoModel::types() const
 }
 
 } // namespace Vremenar
+
+#include "moc_MapInfoModel.cpp"
